requirements.txt
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.24
django-bootstrap3==12.1.0
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==3.0.2
djangorestframework==3.11.0
gunicorn==19.9.0
Pillow==7.1.2
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
whitenoise==4.1.3
